Feature to add default Container Registry to Apps (#15)
This feature enables us to add a default container registry credential to all apps that are created. It's disabled by default and enabled by using the flag --container-registry-url with a value like https://foo:bar@foobar.io. This will add a registry credential and a secret (named azcagit-reg-cred) with foobar.io as server, foo as username and bar as password.